Update patch set 2
Patch Set 2:

> > do we need signal forwarding for darwin given that
 > > darwin doesn't support -linkmode=c-shared?
 > 
 > darwin does support -buildmode=c-archive, which gives C code the
 > chance to register a signal handler before the runtime has started.

I think we've ignored the other possibility: what if other part of the
program registers their signal handler and overwrites ours?
(and even worse, there are issues in the past that some library, for
example, Qt, registers its own signal handlers and also try to be nice.
Its handlers will also run our signal handlers after theirs. The only
problem is that the signal handler is not using SA_ONSTACK, which
not only smashes Go's segmented stack and also invalidated the
assumption in runtime.sigtramp) For iOS, this problem is especially
problematic, as sigaltstack is broken. If our signal handlers
are overwritten, bad thing is almost guaranteed to happen if the
signal is delivered to a Go thread.

For static linking, we can just teach people not to register their own
signal handlers. For iOS, where JIT is forbidden, I doubt there is reason
to do that.

If we tell the user not to register signal handlers of their own, then
we don't need to forward the signals too.

signal forwarding makes sense for android/arm where we must co-exist
with a JVM which is known to register its own signal handlers, but for
darwin, where iOS is the primary user, of -linkmode=c-archive, signal
forwarding isn't that important.

(On android/arm, we can be sure that our shared library is always
loaded after JVM, so the possibility of our signal handlers being
overwritten is negligible unless the user also load another JIT dso.)

Patch-set: 2
Label: Code-Review=0