Update patch set 2
Patch Set 2: -Run-TryBot

> > > do we need signal forwarding for darwin given that
 > > > darwin doesn't support -linkmode=c-shared?
 > >
 > > darwin does support -buildmode=c-archive, which gives C code the
 > > chance to register a signal handler before the runtime has
 > started.
 > 
 > I think we've ignored the other possibility: what if other part of
 > the
 > program registers their signal handler and overwrites ours?

I don't think anyone has ignored that possibility, there have been several issues filed on the subject and it is a traditional problem for managed languages. (For example, see the JVM's libjsig.) It's just not a problem being solved here yet.

I originally started these darwin CLs because I thought iOS's NSException handling was implemented with a signal handler. Most of the way in I realized it's not. It is probably a mach exception handler. I'm cleaning it up and submitting it because if someone decides to use Go with Java in a non-android setting, they will probably want to run on darwin.

Patch-set: 2
Label: Run-TryBot=0