Update patch set 3
Patch Set 3:

Because the signal forwarding can still happen on Go threads,
I don't think it's going to generally work on iOS as we don't
execute signal handlers on a separate stack (sigaltstack doesn't
work on iOS kernels).

I think stack corruption is much worse than not forwarding the
signals.

If you want to do it for Desktop darwin, it's fine, but I don't think
we should do it on iOS.

What we could work is we catch the mach exceptions, determine
if we need to forward, then forward there (it's on a normal thread
stack, so the stack problem above doesn't matter), otherwise
continue to let our signal handler handle it (then we might need
to make the siginfo structs ourselves).

A complete fix for this problem is perhaps to switch from Unix
signals to Mach exception ports on Darwin.

Patch-set: 3
Reviewer: Gerrit User 5055 <5055@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0