Update patch set 2
Patch Set 2: Run-TryBot+1

I'm not sure if this is the best solution to this problem, but it seems to work and makes sense on some level (why *shouldn't* it be okay to ready a goroutine on that P?). An alternative, more targeted solution could be for allocm to call mallocgc directly with a flag that disables funny business like mutator assists (there isn't obviously anything in the P that mallocgc could look at to know it's in a somewhat perilous state).

Patch-set: 2
Label: Run-TryBot=+1