Update patch set 2
Patch Set 2: Run-TryBot+1

The borrowing of P seems to be at the root of this and while I can believe that this solves the current problem as we move to more per-P  constructs borrowing will only become more complicated. It is probably far too late to change this for this cycle but perhaps we can come up with some sort of stress testing that focuses on borrowing.

Patch-set: 2
Label: Run-TryBot=+1