Update patch set 1
Patch Set 1:

(1 comment)

> It bugs me that it's possible for these race tests to be sensitive
 > to the scheduler; like the scheduler is introducing happens-before
 > edges that shouldn't be there. But I don't know enough about the
 > race detector to know if I should be expecting this or not.

In some ways, the race detector is fundamentally sensitive to the actual code paths taken. Consider this ordering:

RLock
Write
RUnlock
Lock
Unlock
RLock
Write
RUnlock

It is just a fact that the first RUnlock happens before Lock, which happens before Unlock, which happens before the second RLock, so the writes are ordered. It's is probably possible to construct slightly more complex scenarios where writing is okay with that kind of synchronization.

The race detector works as well as it does because most orderings do expose problems, but there's no guarantee against something like the above in code that is willfully misusing the primitives.

That said, I do agree that something is amiss in TestRaceRange. I filed #10589 for that.

Patch-set: 1