Update patch set 1
Patch Set 1:

This doesn't feel like it warrants the extra API surface.

Why is this logging at all?

Is it noting bugs in net/http? If so, let's keep the logging but fix the bugs.

If it is just noting bad client behavior, which it seems like is the case in issue 10454, maybe we should just remove the logging.

Patch-set: 1