Update patch set 1
Patch Set 1:

> This doesn't feel like it warrants the extra API surface.

Yeah, see my commit message.
 
 > Why is this logging at all?

On weird behavior from peers, to aid debugging. (your peer is likely under your control too)

 > Is it noting bugs in net/http? If so, let's keep the logging but
 > fix the bugs.

Of course not. If it were bugs I agree we'd just fix them.

 > If it is just noting bad client behavior, which it seems like is
 > the case in issue 10454, maybe we should just remove the logging.

I think the logging is useful to help you fix your client app.

For the http.Server have a dedicated ErrorLog, but down in this cookie-reading code, we don't have access to that.  Maybe we could plumb it through, but I think that might be a lot of layers?

Patch-set: 1