Update patch set 6
Patch Set 6:

I don't understand this failure message.  It looks like linux-amd64-race passed on the dashboard.  Why does this email report failure?  (And it reported failure ~20 minutes before the the dashboard reported success.)  Are there multiple of these builders, and if so why is it stealth?

Patch-set: 6
Label: Code-Review=0