Update patch set 3
Patch Set 3:

> This is kinda gross.

Huh. I hadn't viewed it as gross--just amortizing allocations that happen locally in large numbers--but if you feel strongly about it, I'll happily abandon. Please let me know.


> It seems like you would end up allocating more bytes.

Allocating 128 Regs is ~20k, which is small relative to the rest of the compiler (that's about 50 Nodes), and the allocation is short-lived.

128 Regs is enough for ~75% of the calls to regopt while compiling the stdlib. 64 would be enough for about ~65%. It would also be easy to start smaller and grow.


> Why is the number of allocations important?

I had been working under the theory that removing large numbers of allocations from the compiler will help performance. The call to new(Reg) is about 6% of all allocations.


> Is this a few bytes over some allocator class and actually allocating some unfortunately large number of internally-fragmented bytes?

Nope. Allocator classes are pretty dense for small numbers of bytes, which Regs are.

Patch-set: 3