Update patch set 3
Patch Set 3:

> > Why is the number of allocations important?
 > 
 > I had been working under the theory that removing large numbers of
 > allocations from the compiler will help performance. The call to
 > new(Reg) is about 6% of all allocations.

Not causing GC helps performance, so it's good to avoid unnecessary garbage, but new(A)+new(A) vs new([2]A) should both result in the same number of bytes allocated and thus the same triggering of the GC. So I'd go with the most readable code.

Our allocator shouldn't be slow enough that this would matter. It should have a per-P freelist and available items in its slab from previous allocations.

If you've found this to improve performance, that just suggests to me that the allocator needs work.

Patch-set: 3