Update patch set 1
Patch Set 1:

> This proved to be very subtle in the past.
 > 
 > Did you considered all the additional transitions that this code
 > must handle now? E.g. mcall:
 > 
 > MOVQ	SI, g(CX)	// g = m->g0
 > MOVQ	(g_sched+gobuf_sp)(SI), SP	// sp = m->g0->sched.sp
 > PUSHQ	AX
 > MOVQ	DI, DX
 > MOVQ	0(DI), DI
 > CALL	DI
 > 

The comment for sigprof is still somewhat valid. To switch goroutine, we need to switch the (G, SP) tuple. Because the SP change must happen last, we can check that we're actually running on the stack by checking if the SP is within the Gs stack bounds.

Note that I only jump the stack for systemstack. The other stack changing routines will have mstart at the bottom and not continue any further.

 > How did you test it?

Tested this by running go test -gcflags '-cpuprofile /tmp/cpu' std

Patch-set: 1