cmd/go: `go get` fetches git submodules
`{create,download,tagSync}Cmd` and `tagSyncDefault` have been changed
to string slices, allowing multiple commands to run for each. In
particular, this is used to alter the behaviour of `go get` such that
git submodules are fetched alongside the main repository.

This is a reincarnation of https://codereview.appspot.com/142180043.

Fixes #7764.

Change-Id: I8248efb851130620ef762a765ab8716af430572a