Update patch set 2
Patch Set 2:

> > PTAL
 > 
 > I have the same questions:
 > - are you sure it work?
 > - how did you test?

I tested it by running a program which did nothing cgo callbacks and blocking syscalls in a loop and attaching the http profiler. Besides that, I manually inspected all cases where the SP was set in the runtime to see if they are in one of the functions that I explicitly ignore.

Patch-set: 2