Update patch set 2
Patch Set 2:

The memory model says:

"The closing of a channel happens before a receive that returns a zero value because the channel is closed."

So is this just a bug in the race detector?

I don't want to add gross work-arounds in the code just to "make the race detector happy".

Patch-set: 2