Update patch set 2
Patch Set 2:

> The memory model says:
 > 
 > "The closing of a channel happens before a receive that returns a
 > zero value because the channel is closed."

Happens before a receive, but the close doesn't say anything about what the ordering of the goroutine calling close.

If the client times out immediately, we will call WaitGroup.Wait (which happens in httptest.Server.Close). Then the handler runs and adds one to the WaitGroup. Per http://tip.golang.org/pkg/sync/#WaitGroup.Add, This is not allowed.

The client won't time out immediately though. It will wait for a response for 500ms and by the time it times out, the server will have surely entered the handler. However, the race detector doesn't know this. We have to synchronize by waiting for the Server to enter the handler, which we can do by turning the close into a send.

 > 
 > So is this just a bug in the race detector?
 > 
 > I don't want to add gross work-arounds in the code just to "make
 > the race detector happy".

Patch-set: 2
Label: Code-Review=0