Update patch set 6
Patch Set 6:

> I tried to do this a while back, and it didn't work under certain
 > conditions, so we had to roll it back. I am slightly afraid this CL
 > will fall to the same problems. In particular, I suspect this CL is
 > the root cause for this trybot failure:
 > https://storage.googleapis.com/go-build-log/3edb6fc2/linux-amd64-race_0b92a3c7.log
 > 
 > ok  	runtime/debug	1.027s
 > runtime: invalid pc-encoded table f=type..eq.internal/singleflight.Result
 > pc=0x61fb90 targetpc=0x622bb0 tab=[0/0]0x0
 > value=0 until pc=0x61f9aa
 > value=104 until pc=0x61fb39
 > value=0 until pc=0x61fb3a
 > value=104 until pc=0x61fb4b
 > value=0 until pc=0x61fb4c
 > value=104 until pc=0x61fb5d
 > value=0 until pc=0x61fb5e
 > value=104 until pc=0x61fb7d
 > value=0 until pc=0x61fb7e
 > value=104 until pc=0x61fb90
 > fatal error: invalid runtime symbol table
 > 
 
That's definitely this change. One of the issues I was having was that the linker encoded bogus ranges for external text when linking in internal mode. I avoided it by not invoking the traceback when I was in cgo, but it might be the same case for the race detector.

What link mode does the race detector builds use? Dmitry?

Patch-set: 6