runtime: don't overwrite signal masks for foreign threads
When embedding Go or when foreign threads call into Go,
the existing signal mask is not owned by Go and should not
be overwritten.

Add a parameter to minit(), to track whether the
initializing thread was created by foreign code or not.

This fixes a problem on Android, where the java runtime
ignores certain signals by setting the signal mask for its
initial thread, expecting the mask to be inherited by all
subsequent threads.

Change-Id: Idfd863f930252d09a549bc2f46d28c2a81731770