Update patch set 9
Patch Set 9:

On Wed, May 13, 2015 at 2:23 PM, John C. <jscrockett01@gmail.com> wrote:
> > For x = x[i:i], we know that the resulting cap is zero, and we don't 
> > adjust the base during a slice producing a zero-cap result, 
> > so again base can be ignored completely. 
> 
> Did you mean resulting len is zero (not resulting cap)?

[Please reply using Gerrit. Email replies do not make it into the official review log. I know it's suboptimal.]

No, I meant cap. If the len is zero it might become nonzero during a future slice operation. If the cap is zero, no operation can ever make it nonzero again. Then the only operation influenced by the base pointer is the test for whether it is nil or not. Not advancing the pointer doesn't affect whether it is nil or not, and more importantly it makes sure that we don't advance the pointer past the underlying array into the next object. See the description in https://golang.org/cl/112570044 for additional details.

Patch-set: 9