Update patch set 1
Patch Set 1:

Also, anybody hitting this problem will either have small strings (in which case this won't be a problem) or large strings (in which case they shouldn't be working with strings anyway, and should've stayed in []byte).

But anybody who wants this can write it easily enough themselves. It doesn't pay for its cost having it in standard library.

Patch-set: 1