cmd/internal/gc: rearrange and compact Node fields
* Use a smaller, named type for Top.
  It was int; now uint16.
* Make all Node depths trackers int32.
  Funcdepth was already int32. Make Escloopdepth
  and Decldepth also int32 instead of int.
* Shrink Node.Ostk to uint8.

Then rearrange Node fields to enable better struct packing.
This reduces readability in favor of shrinking
the size of Nodes.

This reduces the size of Node from 368 to 344.
This reduces the memory usage to compile the
rotate tests by about 5.6%.

No functional changes for normal code. Passes toolstash -cmp.

Updates #9933.

Change-Id: I2764c5847fb1635ddc828e2ee385d007d67f03c5