cmd/go, cmd/internal/ld: store information needed for linking to a shared library alongside it, not in it
More from the probably-too-large-for-1.5 files.

The motivation for this is the innocuous looking test case that is added.  This
creates a stack exe -> libdep2.so -> libdep.so -> libruntime.so. The problem
comes from the fact that a function from libdep.so gets inlined all the way
into exe. This (unsurprisingly) means that the object file for exe references
symbols from libdep.so, which means that -ldep needs to be passed when linking
exe and it isn't. The fix is simply to pass it -- there is no harm in passing
it when it's not needed.

The thing is, it's not clear at all in the current code to see how the linker
can know that libdep2 is linked against libdep. It could look through the
DT_NEEDED entries in libdep2 and try to guess which are Go libraries, but it
feels better to be explicit.  So I've implemented an idea that's been knocking
around the back of my head for a while, which is to store all the data that's
needed for linking against a shared library -- that's not accessible from the
list of dynamic symbols -- in a separate file rather than digging it out of
library itself. A bonus of this is that linking against a shared library does
not require the shared library to have not had local symbols stripped out of
it and so I can re-enable the stripping step when making Debian packages and
make libstd.so more than 10 megabytes smaller.

Change-Id: I83d3078d9342e09b72db6eecb11dc9269b401eec