Update patch set 3
Patch Set 3:

(2 comments)

I am not sure whether you wanted a review or not. This looks like debugging code but no actual fix. Note that R=close is undone by uploading a new patch set. If you need to upload not-to-be-reviewed patch sets, please put DO NOT REVIEW in the commit message. Thanks.

Patch-set: 3