Update patch set 1
Patch Set 1:

(3 comments)

'for range s' is working as intended, and the index will advance by one on error, but you need to know the end of the encoding as well as the beginning, so the call to DecodeRune is probably the best you can do.

That is: yes, the discrepancy is known and working as intended.

Patch-set: 1