cmd/link: replace -X name value with -X name=value
People invoking the linker directly already have to change their scripts
to use the new "go tool link", so this is a good time to make the -X flag
behave like all other Go flags and take just a single argument.

I preserved the two-argument behavior during the conversion from C,
but it's non-standard for Go and requires a significant kludge,
now removed.

Change-Id: I91e8df94f4c22b2186e81d7f1016b8767d777eac