Update patch set 1
Patch Set 1:

> I'm not yet convinced a separate file is a great idea.  Doesn't it
 > mean updating all the packaging scripts to push around this extra
 > file along with the shared library?

It's not that bad, at least the way I've done things so far, because for the -dev package I just grab whatever ends up in $GOPATH[0]/pkg after go install (less the actual shared object, which is in a separate package).  So the json just comes along.

 > And where will the extra file live on an installed system?

Currently the above files go to /usr/lib/x86_64-linux-gnu/gocode/pkg.  The .so in the -dev package is a symlink to the file installed by the shlib package into /usr/lib/x86_64-linux-gnu.

 > Also now we have to worry about the format.  Although we require
 > the same version of the go tool, this becomes a useful source of
 > information for people who may not have that exact tool.

Yes, and I admit that's part of the reason that I like this change (I need to access the abi hash during packaging -- reading the value of the local symbol called "go.link.abihashbytes" is also fairly fragile).

Patch-set: 1
Label: Code-Review=0