Update patch set 2
Patch Set 2:

Found one source of lost stack barriers: the setcallerpc used in the select implementation can clobber a stack barrier. I think the cleanest way to fix this is to teach setcallerpc to mark the stack barrier hit if it's about to overwrite it. Alternatively, it could keep the stack barrier and update the saved return PC recorded in the stkbar and let the return go through the stack barrier.

Related to this, getcallerpc needs to know how to recover the original return PC if it sees a stack barrier.

Patch-set: 2