Update patch set 1
Patch Set 1:

It would not work. Simplified, the current code looks like:

 func wait() {
    done = true
    // in release()
    if handle == -1 {
        return
    }
}

 func signal() {
    if handle == -1 {
        return
    }
    if done {
        return
    }
    handle = -1
}

where "done" is read and written atomically but not "handle". Reading done first in signal() is still racy, imagine A calling signal() and B calling wait():

A: if done { ... -> false
B: done = true
B: handle = -1
A: if handle { ...

The handle read and write are racy. Atomic operations on one variable are not enough, I need either a mutex to guarantee the atomicity of reading/writing done and handle, or since there is no logical dependency between both, access handle atomically.

Now, it might be possible to merge handle and done variable, at the cost of extra complexity.

Another variant would be to drop the write to handle and let Win32 API deal with the invalidated handle. Since it was added by the commit introducing the race I discarded this idea.

Patch-set: 1
Label: Code-Review=0