Update patch set 1
Patch Set 1:

It's hard to be happy about this.  It seems error-prone to have a field that must not be changed for OLITERAL but may be changed for other types.  At the least I think it needs a setter function that verifies that we aren't clobbering the Val field for an OLITERAL.

Patch-set: 1
Label: Code-Review=0