Update patch set 1
Patch Set 1:

> > How about changing the official syntax to -X sym=val but silently
 > > accepting -X sym val as well?
 > 
 > I agree that that is our best option at this point.

Why do we need to add -X sym=val syntax? We can't get rid of
-X sym val ever and adding -X sym=val won't simplify the
code anyway.

The reason I think we can't ever get rid of the old syntax is
that it's otherwise impossible for build scripts to support
multiple version of go command without adding conditional
logic. (Even though we drop support for a Go release, there
will still be people using the older Go release, so there will
still be a need for build scripts to support older releases)

I think this kind of breaks is unnecessary trouble for the users.
(it's similar to removing the clean scripts: we know better
solutions exist, but the old one is not broken either.)

Patch-set: 1