Update patch set 2
Patch Set 2:

For Pcrel, it's only used by arm, and it is unfortunately
overloaded:
1. it's used to link constant pool entry back to the
instruction that references it
2. it's used to handle BCASE.

I think we never use BCASE instruction, so we should
be able to get rid of 2.

For 1, we can't get rid of the extra *Prog pointer as long
as we still use constant pool. But I think we can still
use map (not every instruction needs such a pointer)


This change can go in, but in the long term, I'd like
to use this solution.

Patch-set: 2