Update patch set 2
Patch Set 2: Run-TryBot+1

I've stress tested patchset 1 and it is still flaky.
The problem is that the first part of syscallexit (where we decide whether to trace the syscallexit or not) runs completely asynchronously with StartTrace (which emits EvGoIsSyscall or not). So it was possible that StartTrace emits EvGoIsSyscall, but syscallexit decides to not emit EvGoSysExit, or vice versa.

I've added an explicit flag g.sysblocktraced, which says whether StartTrace has emitted EvGoIsSyscall or not. As the result EvGoSysExit is emitted if and only if StartTrace has emitted EvGoIsSyscall about the goroutine.

PTAL

Patch-set: 2
Label: Run-TryBot=+1