Update patch set 1
Patch Set 1: Code-Review-2

Sorry, but not now. Changes are only worth the risk if they are making tangible improvements for end users. This one is at best a complete no-op, at worst introducing a subtle bug that we're all missing. It's a fine change, but it should wait until the dev cycle reopens.

Patch-set: 1
Label: Code-Review=-2