math/big, cmd/compile/internal/big: remove dead code
m was being resliced as the result of looking for the first
non zero word of the mantissa, however m was not used later
in printing.

The code that followed applied the correct trimming of trailing
zeros, so delete the now dead code.

Also apply this to the internal copy of this package in cmd/compile/internal/big

Spotted by Gordon Klass, https://groups.google.com/forum/#!topic/golang-nuts/MdDLbvOjb4o

Change-Id: Ifbebb51ea5e0d86cb8e0422eb184b8634639a733