Update patch set 1
Patch Set 1:

I didn't even look at the implementation, because the standard library isn't accepting new data structures. We can't accept all data structures in the world into the standard library so we have to draw the line somewhere between "all" and "none", and we already have a few very common ones (including hash maps in the core language).  So the question is where do we draw the line?  And my answer is we've already passed it. For existing container types:

heap: we kinda use in the standard library/tools
list: we don't really use in the standard library/tools
ring: likewise

There's no reason this needs to be in the standard library. It doesn't come up very often. It would be better on Github.

Patch-set: 1