Update patch set 1
Patch Set 1:

(1 comment)

> (1 comment)
 > 
 > There's a lot more going on in this CL than the commit message
 > indicates.

FWIW, not a criticism, but there was a lot less going on in this CL before your CL to unembed *Param landed. It may be that Param should be inlined into Name now rather than a separate type, but I think that can wait until the 1.6 cycle.

Patch-set: 1